Truncate Discord messages that exceed 400 chars
Discord has a 400 character limit.
We already dealt with the case of a message exceeding 400 characters when being added to previous messages.
What we didn't deal with are messages that by themselves would exceed 400 characters. It turns out this actually used to break the Discord logging but not anymore with this commit: we now truncate the message before it reaches 400 characters.

We may want to implement a better string splitting logic in the future to avoid truncating the messages but this will have to do for now.
(This also adds some comments to understand why this is here)