Add simple custom commands
This adds simple custom commands. The way it works is that there is a new configuration file where you can define the commands, cooldown and the regex that matches the start of the input. This start will be discarded and only what's after will be taken into account.
Similarly as the Discord configuration, you can specify the path to the configuration file with the ServerManagerBot_CommandsConfPath environment variable.
The commands will be checked every time the process sends new output data, if it finds a corresponding string in the config, it will send that back to the process.